Fills out Page Tests
Goes from empty to the book example, yeehoo. Checks the about me and the home page, and honestly I dropped the content check of the tests because I dont want to inspect pages every single test for that mundane asset. I left the content check hook if I DO want to use it, but better to have it than not.